LCD menu updates including turning extruder on/off
Added ability to turn on/off extrusion from LCD.  Added ability to
change puller multiply ratio from LCD.  Display mm extruded from P axis,
Update main status page.

+++ Marlin/Marlin_main.cpp
@@ -186,6 +186,7 @@ CardReader card;
 #endif
 float homing_feedrate[] = HOMING_FEEDRATE;
 bool axis_relative_modes[] = AXIS_RELATIVE_MODES;
+char extrude_status =0;
 int feedmultiply=100; //100->1 200->2
 int saved_feedmultiply;
 int pullermultiply = 100;
@@ -580,19 +581,19 @@ void loop()
   //FMM generate extruder motion based on LCD inputs
   
   //check that planning buffer is not full
-  
-  //calculate move
-  destination[E_AXIS] = (float)0.1 + current_position[E_AXIS];  //extruder
-  destination[P_AXIS] = (float)0.2 + current_position[P_AXIS]; //puller
-  feedrate=20*60;
-  act_feedrate=feedrate*feedmultiply/60.0/100.0;
-  
-  //send move
-  previous_millis_cmd = millis();  //refresh the kill watchdog timer
-  plan_buffer_line(destination[X_AXIS], destination[Y_AXIS], destination[Z_AXIS], destination[E_AXIS], destination[P_AXIS], act_feedrate, active_extruder);  //FMM added P_AXIS
-  current_position[E_AXIS]=destination[E_AXIS];
-  current_position[P_AXIS]=destination[P_AXIS];
-  
+  if(extrude_status & 1 >0){
+	  //calculate move
+	  destination[E_AXIS] = (float)0.1 + current_position[E_AXIS];  //extruder
+	  destination[P_AXIS] = (float)0.1*pullermultiply/100.0 + current_position[P_AXIS]; //puller
+	  feedrate=20*60;
+	  act_feedrate=feedrate*feedmultiply/60.0/100.0;
+	  
+	  //send move
+	  previous_millis_cmd = millis();  //refresh the kill watchdog timer
+	  plan_buffer_line(destination[X_AXIS], destination[Y_AXIS], destination[Z_AXIS], destination[E_AXIS], destination[P_AXIS], act_feedrate, active_extruder);  //FMM added P_AXIS
+	  current_position[E_AXIS]=destination[E_AXIS];
+	  current_position[P_AXIS]=destination[P_AXIS];
+  }
   
 }
